# docs/openapi.yml
openapi: 3.1.0
info:
  title: Card Collection Price Tracker API
  description: 'REST API for tracking Magic: The Gathering card prices'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
paths:
...
  /collection/trade-settings/binders:
    get:
      tags:
      - trade-settings
      operationId: get_trade_binders
      responses:
        '200':
          description: Binders selected for trade by the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeBindersResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
    post:
      tags:
      - trade-settings
      operationId: add_trade_binder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTradeBinderRequest'
        required: true
      responses:
        '204':
          description: Binder selected for trade
        '400':
          description: Binder name is empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Binder not found in the authenticated user's collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/trade-settings/binders/{name}:
    delete:
      tags:
      - trade-settings
      operationId: remove_trade_binder
      parameters:
      - name: name
        in: path
        description: Binder name
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Binder deselected for trade
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/trade-settings/rarities:
    get:
      tags:
      - trade-settings
      operationId: get_trade_rarities
      responses:
        '200':
          description: Rarities owned within the binders selected for trade, with their trade rule and computed counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RarityFiltersResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
    post:
      tags:
      - trade-settings
      operationId: set_trade_rarity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRarityFilterRequest'
        required: true
      responses:
        '204':
          description: Trade rule updated successfully
        '400':
          description: Invalid rarity code or kept_copies out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/trade-settings/visibility:
    get:
      tags:
      - trade-settings
      operationId: get_trade_visibility
      responses:
        '200':
          description: Current collection visibility
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisibilityResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user has never registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
    put:
      tags:
      - trade-settings
      operationId: set_trade_visibility
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetVisibilityRequest'
        required: true
      responses:
        '204':
          description: Visibility updated successfully
        '400':
          description: Invalid visibility value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user has never registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
...
  /user/{username}:
    get:
      tags:
      - auth
      operationId: get_user_profile
      parameters:
      - name: username
        in: path
        description: Player username
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Public user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Username not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
components:
  schemas:
    AddTradeBinderRequest:
      type: object
      required:
      - binder_name
      properties:
        binder_name:
          type: string
...
    CollectionVisibilityParam:
      type: string
      enum:
      - public
      - trade
      - private
...
    ErrorResponse:
      type: object
      description: |-
        Body of every error response. `error` is a technical message, for diagnosis only; clients
        translate `code` into what they show the user (ADR 0018).
      required:
      - error
      - code
      properties:
        code:
          type: string
          description: Stable snake_case identifier of the error (e.g. `binder_export`, `internal`).
        error:
          type: string
          description: Technical message, in English — never shown to the user as is.
...
    RarityFilterResponse:
      type: object
      required:
      - rarity
      - is_open
      - kept_copies
      - copies
      - proposed
      properties:
        copies:
          type: integer
          format: int64
          minimum: 0
        is_open:
          type: boolean
        kept_copies:
          type: integer
          format: int32
          minimum: 0
        proposed:
          type: integer
          format: int64
          minimum: 0
        rarity:
          type: string
    RarityFiltersResponse:
      type: object
      required:
      - rarities
      properties:
        rarities:
          type: array
          items:
            $ref: '#/components/schemas/RarityFilterResponse'
...
    SetRarityFilterRequest:
      type: object
      required:
      - rarity
      - is_open
      - kept_copies
      properties:
        is_open:
          type: boolean
        kept_copies:
          type: integer
          format: int32
          description: |-
            Signed so out-of-range values (e.g. negative) fail domain validation with a clean 400
            instead of being rejected by the JSON extractor as a 422 before reaching the handler.
        rarity:
          type: string
    SetVisibilityRequest:
      type: object
      required:
      - visibility
      properties:
        visibility:
          $ref: '#/components/schemas/CollectionVisibilityParam'
...
    TradeBindersResponse:
      type: object
      required:
      - binders
      properties:
        binders:
          type: array
          items:
            type: string
...
    UserProfileResponse:
      type: object
      required:
      - id
      properties:
        avatar_url:
          type:
          - string
          - 'null'
        id:
          type: string
        username:
          type:
          - string
          - 'null'
...
    VisibilityResponse:
      type: object
      required:
      - visibility
      properties:
        visibility:
          $ref: '#/components/schemas/CollectionVisibilityParam'
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...